113 Code Cleanup
