+++ Pipfile
@@ -37,7 +37,6 @@ requests-mock = "*"
 werkzeug = "~=0.14.1"
 pytest-camel-collect = "*"
 pytest-cov = "*"
-pytest-sugar = "*"
 
 [requires]
 python_version = "3.7"
