+++ tests/util/__init__.py
@@ -1,1 +1,2 @@
+from .decl import *
 from .monkeypatch import MonkeyPatch
